fix(jaeger): Make jaegercli a thin redirect to internal.NewCommand
