+++ cmd/jaeger/internal/all-in-one.yaml
