+++ cmd/jaeger/jaegercli/command.go
@@ -1,99 +1,20 @@
 // Copyright (c) 2026 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
+// Package jaegercli exposes Jaeger's root CLI command for reuse in custom
+// OpenTelemetry Collector Builder (OCB) distributions, so they retain the
+// embedded all-in-one default config and Jaeger-specific subcommands.
 package jaegercli
 
 import (
-	"embed"
-	"fmt"
-	"log"
-
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/confmap"
-	"go.opentelemetry.io/collector/confmap/provider/envprovider"
-	"go.opentelemetry.io/collector/confmap/provider/fileprovider"
-	"go.opentelemetry.io/collector/confmap/provider/httpprovider"
-	"go.opentelemetry.io/collector/confmap/provider/httpsprovider"
-	"go.opentelemetry.io/collector/confmap/provider/yamlprovider"
 	"go.opentelemetry.io/collector/otelcol"
 
-	"github.com/jaegertracing/jaeger/cmd/internal/docs"
-	"github.com/jaegertracing/jaeger/internal/config"
-	"github.com/jaegertracing/jaeger/internal/storage/v1/elasticsearch/mappings"
-	"github.com/jaegertracing/jaeger/internal/version"
+	"github.com/jaegertracing/jaeger/cmd/jaeger/internal"
 )
 
-//go:embed all-in-one.yaml
-var allInOneYAML embed.FS
-
-const description = "Jaeger backend v2"
-
-// NewCommand builds the root Jaeger cobra command from the given component factories.
-// It wires in the OTel Collector execution engine, the all-in-one default config,
-// and Jaeger-specific subcommands (version, docs, elasticsearch-mappings).
+// NewCommand returns Jaeger's root cobra command wired with the given component
+// factories. It is a thin redirection to internal.NewCommand.
 func NewCommand(factories otelcol.Factories) *cobra.Command {
-	info := component.BuildInfo{
-		Command:     "jaeger",
-		Description: description,
-		Version:     version.Get().GitVersion,
-	}
-
-	settings := otelcol.CollectorSettings{
-		BuildInfo: info,
-		Factories: func() (otelcol.Factories, error) { return factories, nil },
-		ConfigProviderSettings: otelcol.ConfigProviderSettings{
-			ResolverSettings: confmap.ResolverSettings{
-				ProviderFactories: []confmap.ProviderFactory{
-					envprovider.NewFactory(),
-					fileprovider.NewFactory(),
-					httpprovider.NewFactory(),
-					httpsprovider.NewFactory(),
-					yamlprovider.NewFactory(),
-				},
-			},
-		},
-	}
-
-	cmd := otelcol.NewCommand(settings)
-
-	// OTel Collector has no hook for running without a config file, so we intercept
-	// RunE and inject the embedded all-in-one config when no --config flag is set.
-	otelRunE := cmd.RunE
-	cmd.RunE = func(cmd *cobra.Command, args []string) error {
-		return checkConfigAndRun(cmd, args, allInOneYAML.ReadFile, otelRunE)
-	}
-
-	cmd.Short = description
-	cmd.Long = description
-
-	v := viper.New()
-	cmd.AddCommand(version.Command())
-	cmd.AddCommand(docs.Command(v))
-	cmd.AddCommand(mappings.Command())
-	config.AddFlags(v, cmd)
-
-	return cmd
-}
-
-func checkConfigAndRun(
-	cmd *cobra.Command,
-	args []string,
-	getCfg func(name string) ([]byte, error),
-	runE func(cmd *cobra.Command, args []string) error,
-) error {
-	configFlag := cmd.Flag("config")
-	if !configFlag.Changed {
-		log.Print("No '--config' flags detected, using default All-in-One configuration with memory storage.")
-		log.Print("To customize All-in-One behavior, pass a proper configuration.")
-		data, err := getCfg("all-in-one.yaml")
-		if err != nil {
-			return fmt.Errorf("cannot read embedded all-in-one configuration: %w", err)
-		}
-		if err := configFlag.Value.Set("yaml:" + string(data)); err != nil {
-			return fmt.Errorf("cannot apply embedded all-in-one configuration: %w", err)
-		}
-	}
-	return runE(cmd, args)
+	return internal.NewCommand(factories)
 }
